+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [Unreleased]
+- Added github link to modDesc.xml
+- Refactored TempInfo.lua
+- Decreased TempInfo min/max temperatur font size
+
 ## [1.0.0.0] - 2023-04-17
 
 ### Added or Changed